Flytt getSavedRecord til IntegrationTest, rename mockAuthOk, legg til test for arbeidssøker
